moved tri to core math; fixed examples for vec
